Moved TimeServer to Nitrox.Server.Subnautica.Services namespace
