+++ Nitrox.Server.Subnautica/Models/Commands/TimeCommand.cs
@@ -2,6 +2,7 @@ using System.ComponentModel;
 using Nitrox.Model.DataStructures.GameLogic;
 using Nitrox.Server.Subnautica.Models.Commands.Core;
 using Nitrox.Server.Subnautica.Models.GameLogic;
+using Nitrox.Server.Subnautica.Services;
 
 namespace Nitrox.Server.Subnautica.Models.Commands;
