+++ Nitrox.Server.Subnautica/Models/GameLogic/StoryScheduler.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using Nitrox.Model.DataStructures;
 using Nitrox.Model.Subnautica.DataStructures.GameLogic;
 using Nitrox.Server.Subnautica.Models.Packets.Core;
+using Nitrox.Server.Subnautica.Services;
 
 namespace Nitrox.Server.Subnautica.Models.GameLogic
 {
